binance/requirements.txt
py-clob-client
python-dotenv
websockets
certifi
